@@ -23,8 +23,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.connection.foo_connection');
         $this->assertTrue($container->has('old_sound_rabbit_mq.connection_factory.foo_connection'));
         $factory = $container->getDefinition('old_sound_rabbit_mq.connection_factory.foo_connection');
-        $this->assertEquals(array('old_sound_rabbit_mq.connection_factory.foo_connection', 'createConnection'), $definition->getFactory());
-        $this->assertEquals(array(
+        $this->assertEquals(['old_sound_rabbit_mq.connection_factory.foo_connection', 'createConnection'], $definition->getFactory());
+        $this->assertEquals([
             'host' => 'foo_host',
             'port' => 123,
             'user' => 'foo_user',
@@ -33,13 +33,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'lazy' => false,
             'connection_timeout' => 3,
             'read_write_timeout' => 3,
-            'ssl_context' => array(),
+            'ssl_context' => [],
             'keepalive' => false,
             'heartbeat' => 0,
             'use_socket' => false,
             'url' => '',
             'hosts' => [],
-        ), $factory->getArgument(1));
+        ], $factory->getArgument(1));
         $this->assertEquals('%old_sound_rabbit_mq.connection.class%', $definition->getClass());
     }
 
@@ -51,8 +51,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.connection.ssl_connection');
         $this->assertTrue($container->has('old_sound_rabbit_mq.connection_factory.ssl_connection'));
         $factory = $container->getDefinition('old_sound_rabbit_mq.connection_factory.ssl_connection');
-        $this->assertEquals(array('old_sound_rabbit_mq.connection_factory.ssl_connection', 'createConnection'), $definition->getFactory());
-        $this->assertEquals(array(
+        $this->assertEquals(['old_sound_rabbit_mq.connection_factory.ssl_connection', 'createConnection'], $definition->getFactory());
+        $this->assertEquals([
             'host' => 'ssl_host',
             'port' => 123,
             'user' => 'ssl_user',
@@ -61,15 +61,15 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'lazy' => false,
             'connection_timeout' => 3,
             'read_write_timeout' => 3,
-            'ssl_context' => array(
+            'ssl_context' => [
                 'verify_peer' => false,
-            ),
+            ],
             'keepalive' => false,
             'heartbeat' => 0,
             'use_socket' => false,
             'url' => '',
             'hosts' => [],
-        ), $factory->getArgument(1));
+        ], $factory->getArgument(1));
         $this->assertEquals('%old_sound_rabbit_mq.connection.class%', $definition->getClass());
     }
 
@@ -81,8 +81,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.connection.lazy_connection');
         $this->assertTrue($container->has('old_sound_rabbit_mq.connection_factory.lazy_connection'));
         $factory = $container->getDefinition('old_sound_rabbit_mq.connection_factory.lazy_connection');
-        $this->assertEquals(array('old_sound_rabbit_mq.connection_factory.lazy_connection', 'createConnection'), $definition->getFactory());
-        $this->assertEquals(array(
+        $this->assertEquals(['old_sound_rabbit_mq.connection_factory.lazy_connection', 'createConnection'], $definition->getFactory());
+        $this->assertEquals([
             'host' => 'lazy_host',
             'port' => 456,
             'user' => 'lazy_user',
@@ -91,13 +91,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'lazy' => true,
             'connection_timeout' => 3,
             'read_write_timeout' => 3,
-            'ssl_context' => array(),
+            'ssl_context' => [],
             'keepalive' => false,
             'heartbeat' => 0,
             'use_socket' => false,
             'url' => '',
             'hosts' => [],
-        ), $factory->getArgument(1));
+        ], $factory->getArgument(1));
         $this->assertEquals('%old_sound_rabbit_mq.lazy.connection.class%', $definition->getClass());
     }
 
@@ -109,8 +109,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.connection.default');
         $this->assertTrue($container->has('old_sound_rabbit_mq.connection_factory.default'));
         $factory = $container->getDefinition('old_sound_rabbit_mq.connection_factory.default');
-        $this->assertEquals(array('old_sound_rabbit_mq.connection_factory.default', 'createConnection'), $definition->getFactory());
-        $this->assertEquals(array(
+        $this->assertEquals(['old_sound_rabbit_mq.connection_factory.default', 'createConnection'], $definition->getFactory());
+        $this->assertEquals([
             'host' => 'localhost',
             'port' => 5672,
             'user' => 'guest',
@@ -119,13 +119,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'lazy' => false,
             'connection_timeout' => 3,
             'read_write_timeout' => 3,
-            'ssl_context' => array(),
+            'ssl_context' => [],
             'keepalive' => false,
             'heartbeat' => 0,
             'use_socket' => false,
             'url' => '',
             'hosts' => [],
-        ), $factory->getArgument(1));
+        ], $factory->getArgument(1));
         $this->assertEquals('%old_sound_rabbit_mq.connection.class%', $definition->getClass());
     }
 
@@ -164,7 +164,7 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                     'user' => 'cluster_user',
                     'password' => 'cluster_password',
                     'vhost' => '/cluster',
-                    'url' => ''
+                    'url' => '',
                 ],
                 [
                     'host' => 'localhost',
@@ -172,8 +172,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                     'user' => 'guest',
                     'password' => 'guest',
                     'vhost' => '/',
-                    'url' => 'amqp://cluster_url_host:cluster_url_pass@host:10000/cluster_url_vhost'
-                ]
+                    'url' => 'amqp://cluster_url_host:cluster_url_pass@host:10000/cluster_url_vhost',
+                ],
             ],
             'host' => 'localhost',
             'port' => 5672,
@@ -183,7 +183,7 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'lazy' => false,
             'connection_timeout' => 3,
             'read_write_timeout' => 3,
-            'ssl_context' => array(),
+            'ssl_context' => [],
             'keepalive' => false,
             'heartbeat' => 0,
             'use_socket' => false,
@@ -195,7 +195,7 @@ class OldSoundRabbitMqExtensionTest extends TestCase
     public function testFooBinding()
     {
         $container = $this->getContainer('test.yml');
-        $binding = array(
+        $binding = [
             'arguments'                 => null,
             'class'                     => '%old_sound_rabbit_mq.binding.class%',
             'connection'                => 'default',
@@ -204,7 +204,7 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'destination_is_exchange'   => false,
             'nowait'                    => false,
             'routing_key'               => 'baz',
-        );
+        ];
         ksort($binding);
         $key = md5(json_encode($binding));
         $name = sprintf('old_sound_rabbit_mq.binding.%s', $key);
@@ -217,8 +217,8 @@ class OldSoundRabbitMqExtensionTest extends TestCase
     public function testMooBinding()
     {
         $container = $this->getContainer('test.yml');
-        $binding = array(
-            'arguments'                 => array('moo' => 'cow'),
+        $binding = [
+            'arguments'                 => ['moo' => 'cow'],
             'class'                     => '%old_sound_rabbit_mq.binding.class%',
             'connection'                => 'default2',
             'exchange'                  => 'moo',
@@ -226,7 +226,7 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             'destination_is_exchange'   => true,
             'nowait'                    => true,
             'routing_key'               => null,
-        );
+        ];
         ksort($binding);
         $key = md5(json_encode($binding));
         $name = sprintf('old_sound_rabbit_mq.binding.%s', $key);
@@ -238,44 +238,45 @@ class OldSoundRabbitMqExtensionTest extends TestCase
 
     protected function assertBindingMethodCalls(Definition $definition, $binding)
     {
-        $this->assertEquals(array(
-            array(
+        $this->assertEquals(
+            [
+            [
                 'setArguments',
-                array(
-                    $binding['arguments']
-                )
-            ),
-            array(
+                [
+                    $binding['arguments'],
+                ],
+            ],
+            [
                 'setDestination',
-                array(
-                    $binding['destination']
-                )
-            ),
-            array(
+                [
+                    $binding['destination'],
+                ],
+            ],
+            [
                 'setDestinationIsExchange',
-                array(
-                    $binding['destination_is_exchange']
-                )
-            ),
-            array(
+                [
+                    $binding['destination_is_exchange'],
+                ],
+            ],
+            [
                 'setExchange',
-                array(
-                    $binding['exchange']
-                )
-            ),
-            array(
+                [
+                    $binding['exchange'],
+                ],
+            ],
+            [
                 'isNowait',
-                array(
-                    $binding['nowait']
-                )
-            ),
-            array(
+                [
+                    $binding['nowait'],
+                ],
+            ],
+            [
                 'setRoutingKey',
-                array(
-                    $binding['routing_key']
-                )
-            ),
-        ),
+                [
+                    $binding['routing_key'],
+                ],
+            ],
+        ],
             $definition->getMethodCalls()
         );
     }
@@ -287,11 +288,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.foo_producer_producer');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.foo_connection');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.foo_producer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'foo_exchange',
                             'type'        => 'direct',
                             'passive'     => true,
@@ -302,42 +304,42 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setQueueOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => '',
                             'declare'     => false,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setDefaultRoutingKey',
-                    array('')
-                ),
-                array(
+                    [''],
+                ],
+                [
                     'setContentType',
-                    array('text/plain')
-                ),
-                array(
+                    ['text/plain'],
+                ],
+                [
                     'setDeliveryMode',
-                    array(2)
-                ),
-                array(
+                    [2],
+                ],
+                [
                     'setValidator',
-                    array(
+                    [
                         "OldSound\RabbitMqBundle\RabbitMq\JsonValidator",
                         "OldSound\RabbitMqBundle\TestValidation\schema\JsonValidation.schema",
-                        array(
-                            "definitions" => null, 
+                        [
+                            "definitions" => null,
                             "schema_url" => null
-                        )
-                    )
-                )
-            ),
+                        ]
+                    ],
+                ]
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('My\Foo\Producer', $definition->getClass());
@@ -354,16 +356,16 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         }
 
         // test expected aliases
-        $expectedAliases = array(
+        $expectedAliases = [
             ProducerInterface::class . ' $fooProducer' => 'old_sound_rabbit_mq.foo_producer_producer',
             'My\Foo\Producer $fooProducer' => 'old_sound_rabbit_mq.foo_producer_producer',
             ProducerInterface::class . ' $fooProducerAliasedProducer' => 'old_sound_rabbit_mq.foo_producer_aliased_producer',
             'My\Foo\Producer $fooProducerAliasedProducer' => 'old_sound_rabbit_mq.foo_producer_aliased_producer',
             ProducerInterface::class . ' $defaultProducer' => 'old_sound_rabbit_mq.default_producer_producer',
             '%old_sound_rabbit_mq.producer.class% $defaultProducer' => 'old_sound_rabbit_mq.default_producer_producer',
-        );
+        ];
 
-        foreach($expectedAliases as $id => $target) {
+        foreach ($expectedAliases as $id => $target) {
             $this->assertTrue($container->hasAlias($id), sprintf('Container should have %s alias for autowiring support.', $id));
 
             $alias = $container->getAlias($id);
@@ -391,11 +393,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.default_producer_producer');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.default_producer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'default_exchange',
                             'type'        => 'direct',
                             'passive'     => false,
@@ -406,31 +409,31 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setQueueOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => '',
                             'declare'     => false,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setDefaultRoutingKey',
-                    array('')
-                ),
-                array(
+                    [''],
+                ],
+                [
                     'setContentType',
-                    array('text/plain')
-                ),
-                array(
+                    ['text/plain'],
+                ],
+                [
                     'setDeliveryMode',
-                    array(2)
-                )
-            ),
+                    [2],
+                ],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.producer.class%', $definition->getClass());
@@ -444,11 +447,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.foo_consumer_consumer');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.foo_connection');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.foo_consumer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'foo_exchange',
                             'type'        => 'direct',
                             'passive'     => true,
@@ -459,13 +463,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setQueueOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'         => 'foo_queue',
                             'passive'      => true,
                             'durable'      => false,
@@ -474,20 +478,20 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'nowait'       => true,
                             'arguments'    => null,
                             'ticket'       => null,
-                            'routing_keys' => array('android.#.upload', 'iphone.upload'),
+                            'routing_keys' => ['android.#.upload', 'iphone.upload'],
                             'declare'      => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setCallback',
-                    array(array(new Reference('foo.callback'), 'execute'))
-                ),
-                array(
+                    [[new Reference('foo.callback'), 'execute']],
+                ],
+                [
                     'setTimeoutWait',
-                    array(3)
-                )
-            ),
+                    [3],
+                ],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.consumer.class%', $definition->getClass());
@@ -503,15 +507,15 @@ class OldSoundRabbitMqExtensionTest extends TestCase
             return;
         }
 
-        $expectedAliases = array(
+        $expectedAliases = [
             ConsumerInterface::class . ' $fooConsumer' => 'old_sound_rabbit_mq.foo_consumer_consumer',
             '%old_sound_rabbit_mq.consumer.class% $fooConsumer' => 'old_sound_rabbit_mq.foo_consumer_consumer',
             ConsumerInterface::class . ' $defaultConsumer' => 'old_sound_rabbit_mq.default_consumer_consumer',
             '%old_sound_rabbit_mq.consumer.class% $defaultConsumer' => 'old_sound_rabbit_mq.default_consumer_consumer',
             ConsumerInterface::class . ' $qosTestConsumer' => 'old_sound_rabbit_mq.qos_test_consumer_consumer',
-            '%old_sound_rabbit_mq.consumer.class% $qosTestConsumer' => 'old_sound_rabbit_mq.qos_test_consumer_consumer'
-        );
-        foreach($expectedAliases as $id => $target) {
+            '%old_sound_rabbit_mq.consumer.class% $qosTestConsumer' => 'old_sound_rabbit_mq.qos_test_consumer_consumer',
+        ];
+        foreach ($expectedAliases as $id => $target) {
             $this->assertTrue($container->hasAlias($id), sprintf('Container should have %s alias for autowiring support.', $id));
 
             $alias = $container->getAlias($id);
@@ -528,11 +532,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.default_consumer_consumer');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.default_consumer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'default_exchange',
                             'type'        => 'direct',
                             'passive'     => false,
@@ -543,13 +548,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setQueueOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'default_queue',
                             'passive'     => false,
                             'durable'     => true,
@@ -558,16 +563,16 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'nowait'      => false,
                             'arguments'   => null,
                             'ticket'      => null,
-                            'routing_keys' => array(),
+                            'routing_keys' => [],
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setCallback',
-                    array(array(new Reference('default.callback'), 'execute'))
-                )
-            ),
+                    [[new Reference('default.callback'), 'execute']],
+                ],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.consumer.class%', $definition->getClass());
@@ -590,11 +595,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
 
         $this->assertIsArray($setQosParameters);
         $this->assertEquals(
-            array(
+            [
                 1024,
                 1,
-                true
-            ),
+                true,
+            ],
             $setQosParameters
         );
     }
@@ -605,11 +610,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
 
         $this->assertTrue($container->has('old_sound_rabbit_mq.multi_test_consumer_multiple'));
         $definition = $container->getDefinition('old_sound_rabbit_mq.multi_test_consumer_multiple');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'foo_multiple_exchange',
                             'type'        => 'direct',
                             'passive'     => false,
@@ -620,14 +626,14 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setQueues',
-                    array(
-                        array(
-                            'multi_test_1' => array(
+                    [
+                        [
+                            'multi_test_1' => [
                                 'name'         => 'multi_test_1',
                                 'passive'      => false,
                                 'durable'      => true,
@@ -636,11 +642,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                                 'nowait'       => false,
                                 'arguments'    => null,
                                 'ticket'       => null,
-                                'routing_keys' => array(),
-                                'callback'     => array(new Reference('foo.multiple_test1.callback'), 'execute'),
+                                'routing_keys' => [],
+                                'callback'     => [new Reference('foo.multiple_test1.callback'), 'execute'],
                                 'declare'      => true,
-                            ),
-                            'foo_bar_2' => array(
+                            ],
+                            'foo_bar_2' => [
                                 'name'         => 'foo_bar_2',
                                 'passive'      => true,
                                 'durable'      => false,
@@ -649,27 +655,27 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                                 'nowait'       => true,
                                 'arguments'    => null,
                                 'ticket'       => null,
-                                'routing_keys' => array(
+                                'routing_keys' => [
                                     'android.upload',
-                                    'iphone.upload'
-                                ),
-                                'callback'     => array(new Reference('foo.multiple_test2.callback'), 'execute'),
+                                    'iphone.upload',
+                                ],
+                                'callback'     => [new Reference('foo.multiple_test2.callback'), 'execute'],
                                 'declare'      => true,
-                            )
-                        )
-                    )
-                ),
-                array(
+                            ],
+                        ],
+                    ],
+                ],
+                [
                     'setQueuesProvider',
-                    array(
-                        new Reference('foo.queues_provider')
-                    )
-                ),
-                array(
+                    [
+                        new Reference('foo.queues_provider'),
+                    ],
+                ],
+                [
                     'setTimeoutWait',
-                    array(3)
-                )
-            ),
+                    [3],
+                ],
+            ],
             $definition->getMethodCalls()
         );
     }
@@ -682,11 +688,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $this->assertTrue($container->has('old_sound_rabbit_mq.bar_dyn_consumer_dynamic'));
 
         $definition = $container->getDefinition('old_sound_rabbit_mq.foo_dyn_consumer_dynamic');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                        array(
-                            array(
+                        [
+                            [
                                 'name' => 'foo_dynamic_exchange',
                                 'type' => 'direct',
                                 'passive' => false,
@@ -695,24 +702,24 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                                 'internal' => false,
                                 'nowait' => false,
                                 'declare' => true,
-                                'arguments' => NULL,
-                                'ticket' => NULL,
-                            )
-                        )
-                ),
-                array(
+                                'arguments' => null,
+                                'ticket' => null,
+                            ],
+                        ],
+                ],
+                [
                     'setCallback',
-                        array(
-                            array(new Reference('foo.dynamic.callback'), 'execute')
-                        )
-                ),
-                array(
+                        [
+                            [new Reference('foo.dynamic.callback'), 'execute'],
+                        ],
+                ],
+                [
                     'setQueueOptionsProvider',
-                        array(
-                            new Reference('foo.dynamic.provider')
-                        )
-                )
-            ),
+                        [
+                            new Reference('foo.dynamic.provider'),
+                        ],
+                ],
+            ],
             $definition->getMethodCalls()
         );
     }
@@ -725,11 +732,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.foo_anon_consumer_anon');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.foo_connection');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.foo_anon_consumer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'foo_anon_exchange',
                             'type'        => 'direct',
                             'passive'     => true,
@@ -740,14 +748,14 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setCallback',
-                    array(array(new Reference('foo_anon.callback'), 'execute'))
-                )
-            ),
+                    [[new Reference('foo_anon.callback'), 'execute']],
+                ],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.anon_consumer.class%', $definition->getClass());
@@ -761,11 +769,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.default_anon_consumer_anon');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.default_anon_consumer');
-        $this->assertEquals(array(
-                array(
+        $this->assertEquals(
+            [
+                [
                     'setExchangeOptions',
-                    array(
-                        array(
+                    [
+                        [
                             'name'        => 'default_anon_exchange',
                             'type'        => 'direct',
                             'passive'     => false,
@@ -776,14 +785,14 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                             'arguments'   => null,
                             'ticket'      => null,
                             'declare'     => true,
-                        )
-                    )
-                ),
-                array(
+                        ],
+                    ],
+                ],
+                [
                     'setCallback',
-                    array(array(new Reference('default_anon.callback'), 'execute'))
-                )
-            ),
+                    [[new Reference('default_anon.callback'), 'execute']],
+                ],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.anon_consumer.class%', $definition->getClass());
@@ -798,11 +807,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.foo_connection');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.foo_client');
         $this->assertEquals(
-            array(
-                array('initClient', array(true)),
-                array('setUnserializer', array('json_decode')),
-                array('setDirectReplyTo', array(true)),
-            ),
+            [
+                ['initClient', [true]],
+                ['setUnserializer', ['json_decode']],
+                ['setDirectReplyTo', [true]],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.rpc_client.class%', $definition->getClass());
@@ -817,11 +826,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.default_client');
         $this->assertEquals(
-            array(
-                array('initClient', array(true)),
-                array('setUnserializer', array('unserialize')),
-                array('setDirectReplyTo', array(false)),
-            ),
+            [
+                ['initClient', [true]],
+                ['setUnserializer', ['unserialize']],
+                ['setDirectReplyTo', [false]],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertFalse($definition->isLazy());
@@ -837,11 +846,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.lazy_client');
         $this->assertEquals(
-            array(
-                array('initClient', array(true)),
-                array('setUnserializer', array('unserialize')),
-                array('setDirectReplyTo', array(false)),
-            ),
+            [
+                ['initClient', [true]],
+                ['setUnserializer', ['unserialize']],
+                ['setDirectReplyTo', [false]],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertTrue($definition->isLazy());
@@ -856,11 +865,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.foo_server_server');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.foo_connection');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.foo_server');
-        $this->assertEquals(array(
-                array('initServer', array('foo_server')),
-                array('setCallback', array(array(new Reference('foo_server.callback'), 'execute'))),
-                array('setSerializer', array('json_encode')),
-            ),
+        $this->assertEquals(
+            [
+                ['initServer', ['foo_server']],
+                ['setCallback', [[new Reference('foo_server.callback'), 'execute']]],
+                ['setSerializer', ['json_encode']],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.rpc_server.class%', $definition->getClass());
@@ -874,11 +884,12 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.default_server_server');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.default_server');
-        $this->assertEquals(array(
-                array('initServer', array('default_server')),
-                array('setCallback', array(array(new Reference('default_server.callback'), 'execute'))),
-                array('setSerializer', array('serialize')),
-            ),
+        $this->assertEquals(
+            [
+                ['initServer', ['default_server']],
+                ['setCallback', [[new Reference('default_server.callback'), 'execute']]],
+                ['setSerializer', ['serialize']],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.rpc_server.class%', $definition->getClass());
@@ -892,10 +903,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.server_with_queue_options_server');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.server_with_queue_options');
-        $this->assertEquals(array(
-                array('initServer', array('server_with_queue_options')),
-                array('setCallback', array(array(new Reference('server_with_queue_options.callback'), 'execute'))),
-                array('setQueueOptions', array(array(
+        $this->assertEquals(
+            [
+                ['initServer', ['server_with_queue_options']],
+                ['setCallback', [[new Reference('server_with_queue_options.callback'), 'execute']]],
+                ['setQueueOptions', [[
                     'name'         => 'server_with_queue_options-queue',
                     'passive'      => false,
                     'durable'      => true,
@@ -904,11 +916,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                     'nowait'       => false,
                     'arguments'    => null,
                     'ticket'       => null,
-                    'routing_keys' => array(),
+                    'routing_keys' => [],
                     'declare'      => true,
-                ))),
-                array('setSerializer', array('serialize')),
-            ),
+                ]]],
+                ['setSerializer', ['serialize']],
+            ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.rpc_server.class%', $definition->getClass());
@@ -922,10 +934,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $definition = $container->getDefinition('old_sound_rabbit_mq.server_with_exchange_options_server');
         $this->assertEquals((string) $definition->getArgument(0), 'old_sound_rabbit_mq.connection.default');
         $this->assertEquals((string) $definition->getArgument(1), 'old_sound_rabbit_mq.channel.server_with_exchange_options');
-        $this->assertEquals(array(
-            array('initServer', array('server_with_exchange_options')),
-            array('setCallback', array(array(new Reference('server_with_exchange_options.callback'), 'execute'))),
-            array('setExchangeOptions', array(array(
+        $this->assertEquals(
+            [
+            ['initServer', ['server_with_exchange_options']],
+            ['setCallback', [[new Reference('server_with_exchange_options.callback'), 'execute']]],
+            ['setExchangeOptions', [[
                 'name'         => 'exchange',
                 'type'         => 'topic',
                 'passive'      => false,
@@ -936,9 +949,9 @@ class OldSoundRabbitMqExtensionTest extends TestCase
                 'declare'      => true,
                 'arguments'    => null,
                 'ticket'       => null,
-            ))),
-            array('setSerializer', array('serialize')),
-        ),
+            ]]],
+            ['setSerializer', ['serialize']],
+        ],
             $definition->getMethodCalls()
         );
         $this->assertEquals('%old_sound_rabbit_mq.rpc_server.class%', $definition->getClass());
@@ -951,10 +964,11 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $this->assertTrue($container->has('old_sound_rabbit_mq.data_collector'));
         $definition = $container->getDefinition('old_sound_rabbit_mq.data_collector');
 
-        $this->assertEquals(array(
+        $this->assertEquals(
+            [
                 new Reference('old_sound_rabbit_mq.channel.default_producer'),
                 new Reference('old_sound_rabbit_mq.channel.default_consumer'),
-            ),
+            ],
             $definition->getArgument(0)
         );
     }
@@ -979,13 +993,13 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $calls = $definition->getMethodCalls();
         $this->assertEquals('setExchangeOptions', $calls[0][0]);
         $options = $calls[0][1];
-        $this->assertEquals(array('name' => 'bar'), $options[0]['arguments']);
+        $this->assertEquals(['name' => 'bar'], $options[0]['arguments']);
 
         $definition = $container->getDefinition('old_sound_rabbit_mq.consumer_consumer');
         $calls = $definition->getMethodCalls();
         $this->assertEquals('setExchangeOptions', $calls[0][0]);
         $options = $calls[0][1];
-        $this->assertEquals(array('name' => 'bar'), $options[0]['arguments']);
+        $this->assertEquals(['name' => 'bar'], $options[0]['arguments']);
     }
 
     public function testProducerWithoutExplicitExchangeOptionsConnectsToAMQPDefault()
@@ -1008,21 +1022,22 @@ class OldSoundRabbitMqExtensionTest extends TestCase
         $container = $this->getContainer('config_with_enable_logger.yml');
         $definition = $container->getDefinition('old_sound_rabbit_mq.default_consumer_consumer');
         $this->assertTrue(
-            $definition->hasTag('monolog.logger'), 'service should be marked for logger'
+            $definition->hasTag('monolog.logger'),
+            'service should be marked for logger'
         );
     }
 
     private function getContainer($file, $debug = false)
     {
-        $container = new ContainerBuilder(new ParameterBag(array('kernel.debug' => $debug)));
+        $container = new ContainerBuilder(new ParameterBag(['kernel.debug' => $debug]));
         $container->registerExtension(new OldSoundRabbitMqExtension());
 
         $locator = new FileLocator(__DIR__.'/Fixtures');
         $loader = new YamlFileLoader($container, $locator);
         $loader->load($file);
 
-        $container->getCompilerPassConfig()->setOptimizationPasses(array());
-        $container->getCompilerPassConfig()->setRemovingPasses(array());
+        $container->getCompilerPassConfig()->setOptimizationPasses([]);
+        $container->getCompilerPassConfig()->setRemovingPasses([]);
         $container->compile();
 
         return $container;
